test: add int test for AccountDetailsNotVerifiedController and refactor unit test into test folder
